rename(): domain, exception 패키지 형성. exception 관련 클래스명 변경
