+++ src/main/java/lotto/domain/Bonus.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain;
+
+import lotto.exception.ExceptionMessage;
 
 public class Bonus {
     private final int bonusNumber;
@@ -10,7 +12,7 @@ public class Bonus {
 
     private void validateRange(int bonusNumber) {
         if (bonusNumber > 45 || bonusNumber < 1) {
-            throw new IllegalArgumentException(ErrorMessage.OUT_OF_RANGE_NUMBER.getErrorDescription());
+            throw new IllegalArgumentException(ExceptionMessage.OUT_OF_RANGE_NUMBER.getErrorDescription());
         }
     }
