+++ src/test/java/lotto/LottoGameTest.java
@@ -2,6 +2,8 @@ package lotto;
 
 import java.util.HashSet;
 import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.LottoGame;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
